move: CommunityComment 위치
